## Description

This is a work-in-progress implementation of an alternative connection pool implementation.

It takes a lot of inspiration from [HikariCP](https://github.com/brettwooldridge/HikariCP), with slight deviations to better support the specific requirements of Vitess (e.g. the tracking of the connection settings).

The main difference from the current pool is that instead of using various stacks to track connections of different setting types, it tracks all connections that are managed by the pool in a single slice, and uses atomics on the `Pooled[C]` struct to manage the state a connection is in (e.g. whether it's idle or in use).

It seems like just looping over the connections list is faster than managing the connection stacks.

When a connection is requested from the pool, we first try to find an idle connection with matching settings, then fall back to find any idle connection for which we can change the settings, or add the connection to the waitlist (which is reused from the current pool implementation).

When a connection is returned to the pool, we mark the connection as unused and try to hand it over to any of the waiting checkout requests, if there's any.

Another difference to the current pool implementation is that connections that are marked as idle and attempted to be handed over to waiting requests can be "stolen" by concurrent checkout calls. It also does not block on opening a new connection. If we haven't hit the connection limit yet, new connections are opened in the background and if a matching connection is returned to the pool in the meantime, the returned connection will be used instead of waiting on the `Open` operation to finish.

### TODOs

* `Close()` is not implemented correctly - it currently does not close any of the background workers, and I'd like it to be an intermediate operation (orphaning all the in-use connections instead of waiting for connections to be returned to the pool as is done by the current implementation).
* max connection lifetime is not implemented yet.

### Benchmark Results

I added two benchmarks in `fast_pool_test.go`:

```
Histogram:   avg 18.2µs;  min 41ns;  p50 125ns;  max 9.1ms;
  p90 542ns;  p99 2.08µs;  p999 6.58ms;  p9999 9.1ms;
       41ns [3360] ████████████████████████████████████████
      500µs [   0]
        1ms [   0]
      1.5ms [   7]
        2ms [   0]
      2.5ms [   0]
        3ms [   0]
      3.5ms [   0]
        4ms [   4]
      4.5ms [   0]
        5ms [   0]
      5.5ms [   0]
        6ms [   0]
      6.5ms [   2]
        7ms [   0]
      7.5ms [   0]
        8ms [   0]
      8.5ms [   0]
        9ms [   2]
      9.5ms [   0]

BenchmarkPoolWithDifferentSettings-10        	    3375	    350184 ns/op	         0 open_connections	         8.000 setting_changes	         0 waits	      84 B/op	       1 allocs/op
```

```
Histogram:   avg 1.94µs;  min 41ns;  p50 250ns;  max 656µs;
  p90 1µs;  p99 2.45µs;  p999 552µs;  p9999 656µs;
       41ns [3595] ████████████████████████████████████████
       50µs [   2]
      100µs [   1]
      150µs [   1]
      200µs [   1]
      250µs [   0]
      300µs [   1]
      350µs [   0]
      400µs [   0]
      450µs [   0]
      500µs [   3]
      550µs [   4]
      600µs [   0]
      650µs [   1]
      700µs [   0]
      750µs [   0]
      800µs [   0]
      850µs [   0]
      900µs [   0]
      950µs [   0]

BenchmarkFastPoolWithDifferentSettings-10    	    3609	    375357 ns/op	         0 open_connections	         7.000 setting_changes	        10.00 waits	      77 B/op	       0 allocs/op
```

In this benchmark, the new pool shows less variance and better timings.

I also updated the existing benchmark in `load_test.go` to run benchmarks against this new pool (labeled "Fast"):

```
BenchmarkGetPut
BenchmarkGetPut/Legacy/x1-cap64
BenchmarkGetPut/Legacy/x1-cap64-10         	 3959982	       288.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x1-cap64
BenchmarkGetPut/Smart/x1-cap64-10          	 3577706	       380.1 ns/op	      16 B/op	       1 allocs/op
BenchmarkGetPut/Fast/x1-cap64
BenchmarkGetPut/Fast/x1-cap64-10           	 8898646	       133.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x8-cap64
BenchmarkGetPut/Legacy/x8-cap64-10         	 3983781	       286.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x8-cap64
BenchmarkGetPut/Smart/x8-cap64-10          	 3436444	       365.6 ns/op	      16 B/op	       0 allocs/op
BenchmarkGetPut/Fast/x8-cap64
BenchmarkGetPut/Fast/x8-cap64-10           	 8746041	       131.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x32-cap64
BenchmarkGetPut/Legacy/x32-cap64-10        	 3940970	      1227 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x32-cap64
BenchmarkGetPut/Smart/x32-cap64-10         	 1950273	       684.4 ns/op	       8 B/op	       0 allocs/op
BenchmarkGetPut/Fast/x32-cap64
BenchmarkGetPut/Fast/x32-cap64-10          	 9152787	       128.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x128-cap64
BenchmarkGetPut/Legacy/x128-cap64-10       	 2792596	      1240 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x128-cap64
BenchmarkGetPut/Smart/x128-cap64-10        	  922064	      1156 ns/op	       3 B/op	       0 allocs/op
BenchmarkGetPut/Fast/x128-cap64
BenchmarkGetPut/Fast/x128-cap64-10         	 9346239	       120.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x1-cap128
BenchmarkGetPut/Legacy/x1-cap128-10        	 4491429	       282.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x1-cap128
BenchmarkGetPut/Smart/x1-cap128-10         	 3799872	       352.5 ns/op	      16 B/op	       1 allocs/op
BenchmarkGetPut/Fast/x1-cap128
BenchmarkGetPut/Fast/x1-cap128-10          	 8840346	       129.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x8-cap128
BenchmarkGetPut/Legacy/x8-cap128-10        	 4286198	       296.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x8-cap128
BenchmarkGetPut/Smart/x8-cap128-10         	 3368329	       375.4 ns/op	      16 B/op	       1 allocs/op
BenchmarkGetPut/Fast/x8-cap128
BenchmarkGetPut/Fast/x8-cap128-10          	 9240909	       144.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x32-cap128
BenchmarkGetPut/Legacy/x32-cap128-10       	 3802398	       284.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x32-cap128
BenchmarkGetPut/Smart/x32-cap128-10        	 2726973	       524.1 ns/op	      12 B/op	       0 allocs/op
BenchmarkGetPut/Fast/x32-cap128
BenchmarkGetPut/Fast/x32-cap128-10         	 9124257	       127.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x128-cap128
BenchmarkGetPut/Legacy/x128-cap128-10      	 4293358	      1131 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x128-cap128
BenchmarkGetPut/Smart/x128-cap128-10       	 1572543	      1233 ns/op	       2 B/op	       0 allocs/op
BenchmarkGetPut/Fast/x128-cap128
BenchmarkGetPut/Fast/x128-cap128-10        	 8773396	       118.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x1-cap512
BenchmarkGetPut/Legacy/x1-cap512-10        	 4434146	       277.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x1-cap512
BenchmarkGetPut/Smart/x1-cap512-10         	 3653973	       348.1 ns/op	      16 B/op	       1 allocs/op
BenchmarkGetPut/Fast/x1-cap512
BenchmarkGetPut/Fast/x1-cap512-10          	 9064929	       131.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x8-cap512
BenchmarkGetPut/Legacy/x8-cap512-10        	 4377208	       270.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x8-cap512
BenchmarkGetPut/Smart/x8-cap512-10         	 3278119	       374.9 ns/op	      16 B/op	       1 allocs/op
BenchmarkGetPut/Fast/x8-cap512
BenchmarkGetPut/Fast/x8-cap512-10          	 8968564	       136.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x32-cap512
BenchmarkGetPut/Legacy/x32-cap512-10       	 4042659	       279.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x32-cap512
BenchmarkGetPut/Smart/x32-cap512-10        	 3665810	       353.0 ns/op	      16 B/op	       1 allocs/op
BenchmarkGetPut/Fast/x32-cap512
BenchmarkGetPut/Fast/x32-cap512-10         	 9650190	       123.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Legacy/x128-cap512
BenchmarkGetPut/Legacy/x128-cap512-10      	 4571138	       296.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetPut/Smart/x128-cap512
BenchmarkGetPut/Smart/x128-cap512-10       	 2567998	       608.4 ns/op	      11 B/op	       0 allocs/op
BenchmarkGetPut/Fast/x128-cap512
BenchmarkGetPut/Fast/x128-cap512-10        	10234790	       122.2 ns/op	       0 B/op	       0 allocs/op
```

Across the board the new pool seems to perform better.

## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [ ] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [ ] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [ ] Tests were added or are not required
-   [ ] Did the new or modified tests pass consistently locally and on CI?
-   [ ] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->