@arthurschreiber 2025-07-21
I did some more testing, and the performance is not so great in combination with connection settings. The reason for that is that we end up constantly changing the settings on the first few connections in the list if the workload is uncontended but is changing connection settings often. During contention, it's a bit better because we end up with a random connection on which we change the settings, but we still end up with more setting changes than with the connection stack approach. I haven't tried this in a more realistic benchmark, but my expectations that this will perform significantly better are not high.